back/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4

# DB
sqlalchemy==2.0.31
aiomysql==0.2.0
alembic==1.13.2

# Redis
redis[hiredis]==5.0.7

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# HTTP client (토스페이먼츠 API)
httpx==0.27.0

# AWS S3
boto3==1.34.144

# Firebase FCM
firebase-admin==6.5.0

# Utils
python-dateutil==2.9.0

# Dev / Test
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0
factory-boy==3.3.0
